requirements.txt
django-heroku==0.3.1.  
gunicorn==19.9.0 
django-storages==1.7.2
boto3==1.9.250
django-import-export==1.2.0
google-auth==1.6.3
google-cloud-storage==1.20.0
python-dateutil==2.8.0
djangorestframework==3.10.2
psycopg2==2.8.3
psycopg2-binary==2.8.3

